agent-mvp/requirements.txt
chromadb==0.5.23
sentence-transformers==3.0.1
python-dotenv==1.0.0
ollama==0.4.4
